fix(arena): round-2 Copilot review (name-resolution contract + explicit int casts) (#1465)
- ArenaEnemyWeaponCore: document the name-resolution contract on both builders
  — address/value reads are rom-scoped, but NameResolver.GetItemName resolves
  item NAMES via the ambient CoreState.ROM (id-keyed shared cache), so callers
  must keep CoreState.ROM == rom (the editor always opens against the active
  ROM). Matches the reviewer's 'document/enforce' alternative; no churn to the
  shared id-keyed resolver cache.
- Tests: cast computed ROM byte offsets / pointer slots to int at the
  array-index + CopyTo call sites (they already compiled via the long overload
  / widening, but the explicit (int) clears the recurring review nitpick).

Original task: drive PR #1579 (Closes #1465) to merge-ready.

+++ FEBuilderGBA.Avalonia.Tests/ArenaEnemyWeaponRankupParityTests.cs
@@ -58,10 +58,10 @@ public class ArenaEnemyWeaponRankupParityTests : IDisposable
         Assert.True(basicSlot != 0);
         Assert.True(rankupSlot != 0);
 
-        for (uint i = 0; i < 8; i++) bytes[BasicTableAddr + i] = (byte)(i + 1);
-        for (uint i = 0; i < 0x1A; i++) bytes[RankupTableAddr + i] = (byte)(i + 1);
-        BitConverter.GetBytes(BasicTableAddr | 0x08000000u).CopyTo(bytes, basicSlot);
-        BitConverter.GetBytes(RankupTableAddr | 0x08000000u).CopyTo(bytes, rankupSlot);
+        for (int i = 0; i < 8; i++) bytes[(int)BasicTableAddr + i] = (byte)(i + 1);
+        for (int i = 0; i < 0x1A; i++) bytes[(int)RankupTableAddr + i] = (byte)(i + 1);
+        BitConverter.GetBytes(BasicTableAddr | 0x08000000u).CopyTo(bytes, (int)basicSlot);
+        BitConverter.GetBytes(RankupTableAddr | 0x08000000u).CopyTo(bytes, (int)rankupSlot);
 
         rom.LoadLow("arena-1465.gba", bytes, "BE8E01");
         CoreState.ROM = rom;
